64-bit money_t

+++ pg/payment.sql
@@ -1,6 +1,6 @@
 CREATE TABLE payment (
 	user_id INTEGER,
-	posts INTEGER,   -- paid posts in cents
-	logins INTEGER,  -- paid logins in cents
-	awards INTEGER   -- paid awards
+	posts BIGINT,   -- paid posts in cents
+	logins BIGINT,  -- paid logins in cents
+	awards BIGINT   -- paid awards
 );
